requirements/base.txt
Django>=5.2,<6.0
djangorestframework>=3.15,<3.16
djangorestframework-simplejwt>=5.3,<5.4
django-filter>=25.2,<26.0
django-cors-headers>=4.4,<5.0
drf-spectacular>=0.27,<0.28
psycopg2-binary>=2.9,<3.0
python-dotenv>=1.0,<2.0
celery[redis]>=5.6,<5.7
redis>=5.0,<6.0
gunicorn>=22.0,<23.0
